refactor: Create class that wraps height and nth
height와 numberOfPerson를 포장하는 클래스 생성
